ytb-to-tiktok/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytb-to-tiktok"
version = "1.0.0"
description = "Outil pour télécharger des vidéos YouTube et les découper en segments pour TikTok"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Votre Nom", email = "votre.email@example.com"}
]
maintainers = [
    {name = "Votre Nom", email = "votre.email@example.com"}
]
keywords = ["youtube", "video", "download", "split", "tiktok", "social media"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "yt-dlp>=2023.12.30",
    "Pillow>=10.0.0",
    "imageio-ffmpeg>=0.4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
build = [
    "pyinstaller>=5.0.0",
    "auto-py-to-exe>=2.0.0",
]

[project.scripts]
ytb-to-tiktok = "ytb_to_tiktok.cli:main"
ytb-to-tiktok-gui = "gui:main"

[project.urls]
Homepage = "https://github.com/votre-username/ytb-to-tiktok"
Documentation = "https://github.com/votre-username/ytb-to-tiktok#readme"
Repository = "https://github.com/votre-username/ytb-to-tiktok"
"Bug Tracker" = "https://github.com/votre-username/ytb-to-tiktok/issues"

[tool.setuptools]
packages = ["ytb_to_tiktok"]

[tool.setuptools.package-data]
ytb_to_tiktok = ["*.txt", "*.ini", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
